+++ src/_cxxbridge.cpp
@@ -811,7 +811,7 @@ namespace internal {
 #ifndef CXXBRIDGE1_STRUCT_opencompgraph$internal$StreamDataImpl
 #define CXXBRIDGE1_STRUCT_opencompgraph$internal$StreamDataImpl
 struct StreamDataImpl final : public ::rust::Opaque {
-  ::std::size_t get_hash() const noexcept;
+  ::std::uint64_t get_hash() const noexcept;
   const ::rust::Box<::opencompgraph::internal::PixelBlock> &get_pixel_block() const noexcept;
   const ::rust::Box<::opencompgraph::internal::BoundingBox2D> &get_bounding_box() const noexcept;
   const ::rust::Box<::opencompgraph::internal::Matrix4> &get_color_matrix() const noexcept;
@@ -827,7 +827,7 @@ struct NodeImpl final : public ::rust::Opaque {
   ::std::uint8_t get_node_type_id() const noexcept;
   ::opencompgraph::NodeStatus get_status() const noexcept;
   ::std::uint8_t get_status_id() const noexcept;
-  ::std::size_t hash(const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs) noexcept;
+  ::std::uint64_t hash(const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs) noexcept;
   ::opencompgraph::NodeStatus compute(const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs, ::opencompgraph::internal::StreamDataImplShared &output) noexcept;
   ::opencompgraph::AttrState attr_exists(::rust::Str name) const noexcept;
   float get_attr_f32(::rust::Str name) const noexcept;
@@ -872,7 +872,7 @@ namespace internal {
 extern "C" {
 void opencompgraph$internal$cxxbridge1$print_r(const ::opencompgraph::internal::ThingR &r) noexcept;
 
-::std::size_t opencompgraph$internal$cxxbridge1$StreamDataImpl$get_hash(const ::opencompgraph::internal::StreamDataImpl &self) noexcept;
+::std::uint64_t opencompgraph$internal$cxxbridge1$StreamDataImpl$get_hash(const ::opencompgraph::internal::StreamDataImpl &self) noexcept;
 
 const ::rust::Box<::opencompgraph::internal::PixelBlock> *opencompgraph$internal$cxxbridge1$StreamDataImpl$get_pixel_block(const ::opencompgraph::internal::StreamDataImpl &self) noexcept;
 
@@ -892,7 +892,7 @@ const ::rust::Box<::opencompgraph::internal::Matrix4> *opencompgraph$internal$cx
 
 ::std::uint8_t opencompgraph$internal$cxxbridge1$NodeImpl$get_status_id(const ::opencompgraph::internal::NodeImpl &self) noexcept;
 
-::std::size_t opencompgraph$internal$cxxbridge1$NodeImpl$hash(::opencompgraph::internal::NodeImpl &self, const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs) noexcept;
+::std::uint64_t opencompgraph$internal$cxxbridge1$NodeImpl$hash(::opencompgraph::internal::NodeImpl &self, const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs) noexcept;
 
 ::opencompgraph::NodeStatus opencompgraph$internal$cxxbridge1$NodeImpl$compute(::opencompgraph::internal::NodeImpl &self, const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs, ::opencompgraph::internal::StreamDataImplShared &output) noexcept;
 
@@ -943,7 +943,7 @@ void print_r(const ::opencompgraph::internal::ThingR &r) noexcept {
   opencompgraph$internal$cxxbridge1$print_r(r);
 }
 
-::std::size_t StreamDataImpl::get_hash() const noexcept {
+::std::uint64_t StreamDataImpl::get_hash() const noexcept {
   return opencompgraph$internal$cxxbridge1$StreamDataImpl$get_hash(*this);
 }
 
@@ -983,7 +983,7 @@ const ::rust::Box<::opencompgraph::internal::Matrix4> &StreamDataImpl::get_trans
   return opencompgraph$internal$cxxbridge1$NodeImpl$get_status_id(*this);
 }
 
-::std::size_t NodeImpl::hash(const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs) noexcept {
+::std::uint64_t NodeImpl::hash(const ::rust::Vec<::opencompgraph::internal::StreamDataImplShared> &inputs) noexcept {
   return opencompgraph$internal$cxxbridge1$NodeImpl$hash(*this, inputs);
 }
